SpecFixer/requirements.txt
# Core dependencies for SpecFix
pyyaml>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openapi-spec-validator>=0.7.0
prance>=0.22.0
pydantic>=2.0.0
openai>=1.0.0
playwright>=1.40.0

# Development dependencies (optional)
# Install with: pip install -r requirements.txt
# Or for dev: pip install -r requirements.txt -r requirements-dev.txt

